Add interruptible auto-cancel behavior in GitLab CI
This commit introduces the `on_new_commit` trigger to the `auto_cancel` condition in the `.gitlab-ci.yml` file. The change ensures that if a new commit is pushed, the pipeline will be automatically canceled, preventing unnecessary builds. This optimization helps maintain resource utilization and reduces the load on the GitLab CI system.

See merge request swiss-armed-forces/cyber-command/cea/loom!86