Update Libretranslate Dockerfile to include patches
This commit updates the Dockerfile for the libretranslate project. It adds a new directory for patches, copies patch files into this directory, runs a script to apply these patches, and then installs the `gunicorn` package with GUNICORN version 22.0.0. The changes include modifying the installation process to remove the patched files after the operations are completed and ensuring that the health check in `scripts/healthcheck.py` fails on non-200 status codes.

See merge request swiss-armed-forces/cyber-command/cea/loom!93