Optimize Celery worker optimization profile to 'fair'
This change updates the Celery worker command in the Dockerfile to use the 'fair' optimization profile. The 'fair' profile is used here due to the need to ensure that the worker does not reserve multiple messages on the message broker, which can lead to increased delivery counts and potential dead-lettering of messages even if they are never run. This ensures that the worker behaves more predictably and reduces unnecessary overhead.

See merge request swiss-armed-forces/cyber-command/cea/loom!107