Pin Python version to 3.11 in devenv.nix
This change pins the Python version to 3.11 in the `devenv.nix` file, ensuring consistent library resolution for debuggers. This is necessary to maintain compatibility with the Python container images used in the Dockerfiles, and it also simplifies updates by reducing the need to manually update multiple Dockerfiles simultaneously.

See merge request swiss-armed-forces/cyber-command/cea/loom!113