Resolve "Frontend: Prepend new query filed instead of appending"
# 📌 Summary

Besides changing the behavior to prepend instead of append the query parameter, I also fixed a bug with this merge request where when re adding an existing filter, only the filter was removed from the original position and not the white space right after. This lead to white spaces accumulating every time an filter was re clicked.

I fixed that by adjusting the search regex that replaces the filter element from the original position to also select all surrounding white spaces and replaced it with one single line space.

## 🔗 Issue Reference

<!--
If this MR was created to resolve a specific issue, reference it here.
This ensures the issue is automatically closed when the MR is merged.
Example:
Closes #123
-->

Closes #124

See merge request swiss-armed-forces/cyber-command/cea/loom!132

Approved-by: shrewd-laidback palace <shrewd-laidback-palace-736-c41-2c1-e464fc974@swiss-armed-forces-open-source.ch>